// core/config/metadata/type.ts
/* --- Base ------------------------------------------------------------------------------------- */
import type { MetadataRoute } from "next";


export type SiteConfig = {
  Data: {
    url: string;
    appName: string;
    logo: string;
    googleVerification: string;
    twitter: string;
  };
  Theme: {
    light: string;
    dark: string;
  };
  Number: {
    imageWidth: number;
    imageHeight: number;
    logoSize: number;
  };
};

export type SiteLangConfig<T extends string> = {
  [K in T]: {
    Data: {
      name: string;
      locale: string;
      title: string;
      description: string;
      image: string;
      video: string;
      audio: string;
      alt: string;
      shortName: string;
      section: string;
    };
    Check: {
      absoluteTitle: boolean;
    };
    Keywords: string[];
    Tag: string[];
  };
};

export type LanguageConfig<T extends string> = {
  default: T;
  list: Record<T, string>;
};


















/* --- Type Map --------------------------------------------------------------------------------- */
type StringMap = Record<string, string>
type NumberMap = Record<string, number>
type DateMap = Record<string, Date>
type BooleanMap = Record<string, boolean>
type PersonObject = Record<string, {
  "@type": "Person";
  "name": string;
}[]>
/* --- Language --------------------------------------------------------------------------------- */
export type LanguageMap = {
  default: keyof StringMap;
  list: StringMap;
}
export type LanguageObject<StringMap> = Record<string, Record<keyof StringMap, string>>
/* --- Config Site Lang ------------------------------------------------------------------------- */
type ConfigSiteLangMap = {
  Data: StringMap;
  Check: BooleanMap;
  Person: PersonObject;
  Keywords: string[];
  Tag: string[];
}
export type ConfigSiteLangObject<StringMap> = Record<keyof StringMap, ConfigSiteLangMap>
/* --- Config Site ------------------------------------------------------------------------------ */
type ConfigSiteMap = {
  Theme: StringMap;
  Media: StringMap;
  Data: StringMap;
  Date: DateMap;
  Number: NumberMap;
}
export type ConfigSiteObject = ConfigSiteMap

/* --- Config Site Map -------------------------------------------------------------------------- */
/**
 * Icon type for page objects
 * Represents available icon types in the Pelak design system
 */
type Icon = "default" | "none" | "home" | "todo" | "test";
/* --- Page Object Type --------------------------------------------- */
export type PageObjectType = {
  title?: string;
  short?: string;
  description?: string;
  icon?: Icon;
  cover?: string;
  sitemap?: MetadataRoute.Sitemap
}

/* --- Page Map Type ----------------------------------------------- */
export type PageMapType = MetadataRoute.Sitemap;